fix help text
Some of the command line options don't have short names (e.g., `-s` for `--script` but nothing for `--save-all`). This is indicated by a `char` of 0 in the corresponding option in command_line_parser, but this `NUL` still gets printed as an option specifier in the help text (it just looks like a lone dash, `-`). This is clearly misleading (and it messes up the text alignment...). Now the short option section will just be skipped with the appropriate number of spaces.