avoid segfault in single-threaded mode
In src/action/RunAction.cc, there's a reduction step at the end of a run that combines all the TTrees from the worker threads. This is great... if multithreading is enabled. Otherwise, in single-threaded mode, there are no temporary files to open, so trying to clone an empty TChain leads to a segfault.