fix materials
The RPCs are currently simulated as a solid 4.4 cm block of aluminum, minus the 2 mm gas gap. This leads to horribly inaccurate energy losses, much larger than what is realistic considering that the RPC only has ~1 mm total of aluminum thickness. Most of the detector volume is polystyrene foam and the majority of the dense material is the 4 mm of bakelite surrounding the gas.